+++ external/three_addons.js
@@ -1795,7 +1795,6 @@ var OrbitControls = function ( object, domElement ) {
 
   // this method is exposed, but perhaps it would be better if we can make it private...
   this.update = function () {
-
     var offset = new Vector3();
 
     // so camera.up is the orbit axis
