+++ interactives/snaps.css
@@ -138,7 +138,10 @@ h2 {
     overflow-y: hidden;
 }
 
-#drawing-surface {
+#main-surface {
+    position: absolute;
+    left: 0;
+    top: 0;
     height: 100%;
     width: calc(100% - 500px);
 }
