+++ interactives/symmetry_groups.html
@@ -9,7 +9,7 @@
 <body>
 
 <div class="symmetry-groups-wrapper">
-<div id="drawing-surface"></div>
+<div id="main-surface"></div>
 <div class="sidebar">
     <h1>Symmetry Groups</h1>
